[IMP] rest_log: split exception tests
