+++ rest_log/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_db_logging
+from . import test_db_logging_exception
